requirements.txt
Django>=5.1
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
